Added Constraint Inclusions into Parser
Added Constraints Into final check into parser. 

!! Still need to add the LEC 9 should be evening check. That is unwanted statements for all slots for the course that are less than 18 !!

I.e since all 500 courses have to be scheduled into different time slots, all 500 level courses have non-compat statements added to each other at the end of the parse. This was done for all constraints, but the evening. 

All courses have unwanted statements for slots at 11:00-12:30.

The 813 / 913 case was handled, as well as the transitive case : -- This should be tested.... 



Also fixed an issue with pair, that had it so the Hash value could be different for objects that should be equal.